backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
torch>=2.1.0
tiktoken>=0.5.0
transformers>=4.36.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
ddgs